chore: do not run the throttle module at all
